migrate test_telegram to pytest
